requirements.txt
PyYAML>=6.0
pytest>=7.0
maturin>=1.0
streamlit>=1.28
# Optional for autonomous online exploration (DuckDuckGo, no API key):
requests>=2.28
beautifulsoup4>=4.11
duckduckgo-search>=4.0
